Fix utterly broken fs_expand_path function
fs_expand_path was completely broken. Replace with strdup().

In other news, handling endian conversion for stat buffers etc
will be interesting.

meterpreter > ls /etc/passwd
122201/-w------x  3187867648  lin  1969-12-31 19:00:00 -0500  /etc/passwd

I'll document some ideas for a test plan for the posix meterpreter.

It will probably be based upon the metsvc and resource scripts in
order to verify that changes do not break it.