WIP commit for ARM, seems that everything works.
Haven't tested loading ext_server_stdapi.armlelso, but that comes
next.

Added a file for a TESTING plan for when pull requests are issued.
I would like to see some automated testing occur later on so that
various issues are detected quickly and automatically, in case
that's not happening presently.